requirements.txt
Pillow==4.3.0
Flask==0.12
Flask-Compress==1.4.0
Flask-SQLAlchemy==2.2
kafka-python==1.3.3
landmark-extractor==0.1.5.0
beautifulsoup4==4.5.3
phonenumbers==8.0.1
MySQL-python==1.2.5
SQLAlchemy==1.1.5
enum34==1.1.6
requests==2.13.0
chardet==3.0.2
uWSGI==2.0.14
github3.py==0.9.6
tldextract==2.1.0
